requirements.txt
altgraph==0.17
compoundfiles==0.3
compressed-rtf==1.0.6
dnspython==2.0.0
future==0.18.2
IMAPClient==2.1.0
ipwhois==1.1.0
olefile==0.46
pefile==2019.4.18
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pywin32-ctypes==0.2.0
six==1.15.0
tzlocal==1.5.1
